[zeroc-ice] Revise non-windows build

+++ ports/zeroc-ice/portfile.cmake
@@ -4,15 +4,15 @@ vcpkg_from_github(
     REF "v${VERSION}"
     SHA512 07d7c439fbe1f69d808d05a11f32e09cdd8d4df2a93b6f253496304e0a521d417212ae688e316b4450dae406b59d1a460025b51ecd0614c69e48d86c0a6f81c5
     PATCHES
-        mcppd_fix.patch
         no-werror.patch
+        readline.patch
+        rpath-link.patch
         fix-missing-functional.patch
 )
 
 set(RELEASE_TRIPLET ${TARGET_TRIPLET}-rel)
 set(DEBUG_TRIPLET ${TARGET_TRIPLET}-dbg)
 
-set(UNIX_BUILD_DIR "${SOURCE_PATH}")
 set(WIN_DEBUG_BUILD_DIR "${CURRENT_BUILDTREES_DIR}/${DEBUG_TRIPLET}")
 set(WIN_RELEASE_BUILD_DIR "${CURRENT_BUILDTREES_DIR}/${RELEASE_TRIPLET}")
 
@@ -101,7 +101,7 @@ if("icegridtools" IN_LIST FEATURES)
     vcpkg_list(APPEND ICE_OPTIONAL_COMPONENTS_MSBUILD "/t:C++98\\icegridadmin")
     vcpkg_list(APPEND ICE_OPTIONAL_COMPONENTS_MSBUILD "/t:C++98\\icegridregistry")
     vcpkg_list(APPEND ICE_OPTIONAL_COMPONENTS_MSBUILD "/t:C++98\\icegridnode")
-    vcpkg_list(APPEND ICE_OPTIONAL_COMPONENTS_MAKE "icegridnode")
+    vcpkg_list(APPEND ICE_OPTIONAL_COMPONENTS_MAKE "icegridadmin")
     vcpkg_list(APPEND ICE_OPTIONAL_COMPONENTS_MAKE "icegridregistry")
     vcpkg_list(APPEND ICE_OPTIONAL_COMPONENTS_MAKE "icegridnode")
     list(APPEND pkgconfig_packages expat)
@@ -141,126 +141,7 @@ if("icelocatordiscovery" IN_LIST FEATURES)
     vcpkg_list(APPEND ICE_OPTIONAL_COMPONENTS_MAKE "IceLocatorDiscovery")
 endif()
 
-if(NOT VCPKG_TARGET_IS_WINDOWS)
-    # Clean up for the first round (important for install --editable)
-    vcpkg_execute_build_process(
-        COMMAND make distclean
-        WORKING_DIRECTORY ${SOURCE_PATH}/cpp
-        LOGNAME make-clean-${TARGET_TRIPLET}
-    )
-
-    if(EXISTS "${UNIX_BUILD_DIR}/cpp/lib")
-        file(REMOVE_RECURSE "${UNIX_BUILD_DIR}/cpp/lib")
-    endif()
-    if(EXISTS "${UNIX_BUILD_DIR}/cpp/lib64")
-        file(REMOVE_RECURSE "${UNIX_BUILD_DIR}/cpp/lib64")
-    endif()
-    file(REMOVE_RECURSE "${UNIX_BUILD_DIR}/cpp/bin")
-
-    # Setting these as environment variables, as .d files aren't generated
-    # the first time passing them as arguments to make.
-    set(ENV{MCPP_HOME} ${CURRENT_INSTALLED_DIR})
-    set(ENV{EXPAT_HOME} ${CURRENT_INSTALLED_DIR})
-    set(ENV{BZ2_HOME} ${CURRENT_INSTALLED_DIR})
-    set(ENV{LMDB_HOME} ${CURRENT_INSTALLED_DIR})
-    set(ENV{CPPFLAGS} "-I${CURRENT_INSTALLED_DIR}/include")
-    set(ENV{LDFLAGS} "-L${CURRENT_INSTALLED_DIR}/debug/lib")
-
-    set(ICE_BUILD_CONFIG "shared cpp11-shared")
-    if(VCPKG_LIBRARY_LINKAGE STREQUAL "static")
-        set(ICE_BUILD_CONFIG "static cpp11-static")
-    endif()
-    if(NOT VCPKG_BUILD_TYPE)
-    message(STATUS "Building ${TARGET_TRIPLET}-dbg")
-        vcpkg_execute_build_process(
-            COMMAND make
-                V=1
-                "prefix=${CURRENT_PACKAGES_DIR}/debug"
-                linux_id=vcpkg
-                "CONFIGS=${ICE_BUILD_CONFIG}"
-                USR_DIR_INSTALL=yes
-                OPTIMIZE=no
-                ${ICE_OPTIONAL_COMPONENTS_MAKE}
-                "-j${VCPKG_CONCURRENCY}"
-            WORKING_DIRECTORY ${SOURCE_PATH}/cpp
-            LOGNAME make-${TARGET_TRIPLET}-dbg
-        )
-
-        # Install debug libraries to packages directory
-        file(MAKE_DIRECTORY "${CURRENT_PACKAGES_DIR}/debug/lib")
-        if(EXISTS "${UNIX_BUILD_DIR}/cpp/lib64")
-            file(GLOB ICE_DEBUG_LIBRARIES "${UNIX_BUILD_DIR}/cpp/lib64/*")
-        else()
-            file(GLOB ICE_DEBUG_LIBRARIES "${UNIX_BUILD_DIR}/cpp/lib/*")
-        endif()
-        file(COPY ${ICE_DEBUG_LIBRARIES} DESTINATION "${CURRENT_PACKAGES_DIR}/debug/lib")
-
-        # Clean up for the next round
-        vcpkg_execute_build_process(
-            COMMAND make distclean
-            WORKING_DIRECTORY ${SOURCE_PATH}/cpp
-            LOGNAME make-clean-${TARGET_TRIPLET}
-        )
-
-        if(EXISTS "${UNIX_BUILD_DIR}/cpp/lib")
-            file(REMOVE_RECURSE "${UNIX_BUILD_DIR}/cpp/lib")
-        endif()
-        if(EXISTS "${UNIX_BUILD_DIR}/cpp/lib64")
-            file(REMOVE_RECURSE "${UNIX_BUILD_DIR}/cpp/lib64")
-        endif()
-        file(REMOVE_RECURSE "${UNIX_BUILD_DIR}/cpp/bin")
-    endif() # TODO: get-cmake-vars!
-    # Release build
-    set(ENV{LDFLAGS} "-L${CURRENT_INSTALLED_DIR}/lib")
-    message(STATUS "Building ${TARGET_TRIPLET}-rel")
-    vcpkg_execute_build_process(
-        COMMAND make
-            V=1
-            "prefix=${CURRENT_PACKAGES_DIR}"
-            linux_id=vcpkg
-            "CONFIGS=${ICE_BUILD_CONFIG}"
-            USR_DIR_INSTALL=yes
-            OPTIMIZE=yes
-            ${ICE_OPTIONAL_COMPONENTS_MAKE}
-            "-j${VCPKG_CONCURRENCY}"
-        WORKING_DIRECTORY ${SOURCE_PATH}/cpp
-        LOGNAME make-${TARGET_TRIPLET}-rel
-    )
-
-    # Install release libraries and other files to packages directory
-    file(MAKE_DIRECTORY "${CURRENT_PACKAGES_DIR}/lib")
-    file(MAKE_DIRECTORY "${CURRENT_PACKAGES_DIR}/include")
-    file(MAKE_DIRECTORY "${CURRENT_PACKAGES_DIR}/share/ice/slice")
-    file(MAKE_DIRECTORY "${CURRENT_PACKAGES_DIR}/tools/${PORT}")
-
-    install_includes("${UNIX_BUILD_DIR}/cpp/include" "${ICE_INCLUDE_SUB_DIRECTORIES}")
-    install_includes("${UNIX_BUILD_DIR}/cpp/include/generated" "${ICE_INCLUDE_SUB_DIRECTORIES}")
-    install_slices("${SOURCE_PATH}/slice" "${ICE_INCLUDE_SUB_DIRECTORIES}")
-    if(EXISTS "${UNIX_BUILD_DIR}/cpp/lib64")
-        file(GLOB ICE_RELEASE_LIBRARIES "${UNIX_BUILD_DIR}/cpp/lib64/*")
-    else()
-        file(GLOB ICE_RELEASE_LIBRARIES "${UNIX_BUILD_DIR}/cpp/lib/*")
-    endif()
-    file(COPY ${ICE_RELEASE_LIBRARIES} DESTINATION "${CURRENT_PACKAGES_DIR}/lib")
-    file(GLOB ICE_RELEASE_EXECUTABLES "${UNIX_BUILD_DIR}/cpp/bin/*")
-    file(COPY ${ICE_RELEASE_EXECUTABLES} DESTINATION "${CURRENT_PACKAGES_DIR}/tools/${PORT}")
-
-    # Clean up
-    vcpkg_execute_build_process(
-        COMMAND make distclean
-        WORKING_DIRECTORY ${SOURCE_PATH}/cpp
-        LOGNAME make-clean-after-build-${TARGET_TRIPLET}
-    )
-
-    if(EXISTS "${UNIX_BUILD_DIR}/cpp/lib")
-        file(REMOVE_RECURSE "${UNIX_BUILD_DIR}/cpp/lib")
-    endif()
-    if(EXISTS "${UNIX_BUILD_DIR}/cpp/lib64")
-        file(REMOVE_RECURSE "${UNIX_BUILD_DIR}/cpp/lib64")
-    endif()
-    file(REMOVE_RECURSE "${UNIX_BUILD_DIR}/cpp/bin")
-
-else() # VCPKG_TARGET_IS_WINDOWS
+if(VCPKG_TARGET_IS_WINDOWS AND NOT VCPKG_TARGET_IS_MINGW)
 
     # Fix project files to prevent nuget restore of dependencies and
     # remove hard coded runtime linkage
@@ -305,6 +186,46 @@ else() # VCPKG_TARGET_IS_WINDOWS
 
     vcpkg_clean_msbuild()
 
+else()
+
+    file(COPY "${CMAKE_CURRENT_LIST_DIR}/configure" DESTINATION "${SOURCE_PATH}")
+
+    vcpkg_list(SET options)
+    if(VCPKG_CROSSCOMPILING)
+        vcpkg_list(APPEND options
+            "slice2cpp_path=${CURRENT_HOST_INSTALLED_DIR}/tools/zeroc-ice/slice2cpp${VCPKG_HOST_EXECUTABLE_SUFFIX}"
+        )
+    endif()
+
+    list(JOIN ICE_OPTIONAL_COMPONENTS_MAKE " " components)
+    vcpkg_configure_make(
+        SOURCE_PATH "${SOURCE_PATH}"
+        COPY_SOURCE
+        OPTIONS
+            "COMPONENTS=${components}"
+            ${options}
+        OPTIONS_RELEASE
+            OPTIMIZE=yes
+        OPTIONS_DEBUG
+            OPTIMIZE=no
+    )
+    vcpkg_install_make(
+        MAKEFILE "Makefile.vcpkg"
+    )
+
+    string(REPLACE ";icebox;" ";icebox;icebox++11;" tools ";${ICE_OPTIONAL_COMPONENTS_MAKE};")
+    list(FILTER tools INCLUDE REGEX "^ice|^glacier")
+    if(tools)
+        vcpkg_copy_tools(TOOL_NAMES ${tools} AUTO_CLEAN)
+    endif()
+    if(NOT VCPKG_CROSSCOMPILING)
+        vcpkg_copy_tools(TOOL_NAMES slice2cpp SEARCH_DIR "${CURRENT_BUILDTREES_DIR}/${TARGET_TRIPLET}-rel/cpp/bin")
+    endif()
+
+    file(REMOVE_RECURSE
+        "${CURRENT_PACKAGES_DIR}/debug/include"
+        "${CURRENT_PACKAGES_DIR}/debug/share"
+    )
 endif()
 
 # Remove unnecessary static libraries.
@@ -312,7 +233,9 @@ file(GLOB PDLIBS "${CURRENT_PACKAGES_DIR}/debug/lib/*")
 file(GLOB PRLIBS "${CURRENT_PACKAGES_DIR}/lib/*")
 list(FILTER PDLIBS INCLUDE REGEX ".*(([Ii]ce[Uu]til|[Ss]lice)d?\.([a-z]+))$")
 list(FILTER PRLIBS INCLUDE REGEX ".*(([Ii]ce[Uu]til|[Ss]lice)d?\.([a-z]+))$")
-file(REMOVE ${PDLIBS} ${PRLIBS})
+if(NOT "${PDLIBS}${PRLIBS}" STREQUAL "")
+    file(REMOVE ${PDLIBS} ${PRLIBS})
+endif()
 
 file(INSTALL "${CURRENT_PORT_DIR}/vcpkg-ci-IceConfig.cmake" DESTINATION "${CURRENT_PACKAGES_DIR}/share/vcpkg-ci/cmake-user")
 file(INSTALL "${CURRENT_PORT_DIR}/vcpkg-cmake-wrapper.cmake" DESTINATION "${CURRENT_PACKAGES_DIR}/share/ice")
